bugfix: concurrency issue in reader factory storing writer pool stats to table for analysis

+++ core/src/main/java/com/questdb/misc/Unsafe.java
@@ -90,7 +90,7 @@ public final class Unsafe {
     }
 
     public static boolean cas(long[] array, int index, long expected, long value) {
-        return Unsafe.getUnsafe().compareAndSwapLong(array, Unsafe.LONG_OFFSET + (index << Unsafe.LONG_SCALE), expected, value);
+        return Unsafe.getUnsafe().compareAndSwapLong(array, Unsafe.LONG_OFFSET + (((long) index) << Unsafe.LONG_SCALE), expected, value);
     }
 
     public static void free(long ptr, long size) {
